+++ src/app/components/ChatInput/index.tsx
@@ -61,7 +61,7 @@ import XRegExp from 'xregexp';
 import SelectMedia from '../SelectMedia';
 import MediaPreview, {IMediaItem} from '../Uploader';
 import ContactPicker from '../ContactPicker';
-import {C_MESSAGE_TYPE} from '../../repository/message/consts';
+import {C_MESSAGE_TYPE, C_REPLY_ACTION} from '../../repository/message/consts';
 import RiverTime from '../../services/utilities/river_time';
 import Broadcaster from '../../services/broadcaster';
 import MapPicker, {IGeoItem} from '../MapPicker';
@@ -165,6 +165,11 @@ export const C_TYPING_INTERVAL = 5000;
 
 const textBoxClasses: string[] = range(1, 13).map(o => `_${o}-line`);
 
+interface IKeyboardLayout {
+    layout: ReplyKeyboardMarkup.AsObject | undefined;
+    msgId: number;
+}
+
 class ChatInput extends React.Component<IProps, IState> {
     private mentionContainer: any = null;
     private textarea: any = null;
@@ -221,7 +226,8 @@ class ChatInput extends React.Component<IProps, IState> {
     private isMobileBrowser = isMobile();
     private callerId: number = UniqueId.getRandomId();
     private selectMediaRef: SelectMedia | undefined;
-    private botKeyboard: ReplyKeyboardMarkup.AsObject | undefined;
+    private botKeyboard: IKeyboardLayout | undefined;
+    private firstLoad: boolean = true;
 
     constructor(props: IProps) {
         super(props);
@@ -284,6 +290,8 @@ class ChatInput extends React.Component<IProps, IState> {
 
     public setPeer(peer: InputPeer | null) {
         if (peer && this.state.peer !== peer) {
+            this.firstLoad = true;
+            this.botKeyboard = undefined;
             this.preventMessageSend = false;
             if (this.state.voiceMode === 'lock' || this.state.voiceMode === 'down') {
                 this.voiceCancelHandler();
@@ -462,23 +470,43 @@ class ChatInput extends React.Component<IProps, IState> {
         }
     }
 
-    public setBot(isBot: boolean, data: ReplyKeyboardMarkup.AsObject | undefined) {
-        const lastKeyboard = this.botKeyboard;
-        this.botKeyboard = data;
-        let check = false;
-        if (this.state.botKeyboard) {
-            this.setState({
-                isBot,
+    public setBot(isBot: boolean, data: IKeyboardLayout | undefined) {
+        if (this.firstLoad && this.state.peer) {
+            this.firstLoad = false;
+            this.messageRepo.getLastIncomingMessage(this.state.peer.getId() || '').then((msg) => {
+                if (msg) {
+                    if ((!this.botKeyboard || (this.botKeyboard && this.botKeyboard.msgId < (msg.id || 0))) && msg.replymarkup === C_REPLY_ACTION.ReplyKeyboardMarkup) {
+                        this.botKeyboard = {
+                            layout: msg.replydata,
+                            msgId: msg.id || 0
+                        };
+                        this.forceUpdate();
+                    }
+                }
             });
-            check = true;
+        }
+        if ((this.botKeyboard && data && this.botKeyboard.msgId < data.msgId) || !this.botKeyboard) {
+            const lastKeyboard = this.botKeyboard;
+            this.botKeyboard = data;
+            let check = false;
+            if (this.state.botKeyboard) {
+                this.setState({
+                    isBot,
+                });
+                check = true;
+            } else if (this.state.isBot !== isBot) {
+                this.setState({
+                    isBot,
+                });
+                check = true;
+            }
+            if (!check && lastKeyboard !== data) {
+                this.forceUpdate();
+            }
         } else if (this.state.isBot !== isBot) {
             this.setState({
                 isBot,
             });
-            check = true;
-        }
-        if (!check && lastKeyboard !== data) {
-            this.forceUpdate();
         }
     }
 
@@ -2040,22 +2068,22 @@ class ChatInput extends React.Component<IProps, IState> {
     }
 
     private renderBotKeyboard() {
-        if (!this.botKeyboard) {
+        if (!this.botKeyboard || !this.botKeyboard.layout) {
             return '';
         }
-        let height = this.botKeyboard.rowsList.length * 32 + 4;
+        let height = this.botKeyboard.layout.rowsList.length * 32 + 4;
         if (height > 110) {
             height = 110;
             return (<div style={{height: `${height}px`}}>
                 <Scrollbars
                     hideTracksWhenNotNeeded={false}
                     universal={true}>
-                    <BotLayout rows={(this.botKeyboard ? this.botKeyboard.rowsList : undefined)}
+                    <BotLayout rows={(this.botKeyboard ? this.botKeyboard.layout.rowsList : undefined)}
                                prefix="keyboard-bot" onAction={this.props.onBotButtonAction}/>
                 </Scrollbars>
             </div>);
         }
-        return (<BotLayout rows={(this.botKeyboard ? this.botKeyboard.rowsList : undefined)}
+        return (<BotLayout rows={(this.botKeyboard ? this.botKeyboard.layout.rowsList : undefined)}
                            prefix="keyboard-bot" onAction={this.props.onBotButtonAction}/>);
     }